+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/business_result/CongratsViewModel.kt
@@ -14,5 +14,6 @@ internal data class CongratsViewModel(
     val actionCardViewData: MLBusinessActionCardViewData?,
     val crossSellingBoxData: List<MLBusinessCrossSellingBoxData>?,
     val viewReceipt: PaymentCongratsResponse.Action?,
-    val customOrder: Boolean
+    val customOrder: Boolean,
+    val operationInfo: PaymentCongratsResponse.OperationInfo?
 )